requirements.txt
yt-dlp
requests
Pillow
pywebview